requirements.txt
flask
python-dotenv
crewai
crewai[tools]
langchain-perplexity
pandas
requests